requirements.txt
asgiref==3.8.1
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.16
django-cors-headers==4.6.0
django-filter==2.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==2.9
importlib-metadata==1.6.0
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
mysqlclient==2.2.6
pillow==11.0.0
Pygments==2.6.1
PyJWT==2.10.1
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==3.0.1
urllib3==1.25.9
zipp==3.1.0
